backend/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
black==21.10b0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
python-dotenv==0.19.1
pytz==2021.3
regex==2021.11.2
requests==2.26.0
soupsieve==2.3
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==3.10.0.2
urllib3==1.26.7
